requirements.txt
pandas
Pillow
python-dotenv
fastapi[standard]
pre-commit
requests
openai